Reworked The Back for usability.  Moved functions from functions_back into files based on what pages they operate on.
